sierra_adapter/sierra_progress_reporter/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.3.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
idna==2.10
    # via requests
requests==2.25.1
    # via -r requirements.in
urllib3==1.26.5
    # via requests
